Introduce VirtIONetRaw to allow custom NIC buffer management and used buffer notification suppression (#111)
* Introduce `VirtIONetRaw` to allow custom NIC buffer management

* wip: add used buffer notification suppression but do not work

* feat: disable_interrupts by disable_dev_notify of both xmit_queue and recv_queue

* feat: rebase new-netdev to master and support netdev used buffer notification suppression

* fix: cargo fmt

* fix: workflow test bugs

* fix: combine enable_dev_notify and disable_dev_notify

* fix: delete transmit_wait

* feat: rebase to master

* fix: cargo fmt in examples

* feat: add RING_EVENT_IDX feature in net dev

* feat: modify can send, delete can_recv

* Implement can_recv in terms of poll_receive.

* Fix typos.

---------
