+++ examples/x86_64/Makefile
@@ -3,7 +3,7 @@ target := x86_64-unknown-none
 mode := release
 kernel := target/$(target)/$(mode)/$(arch)
 img := target/$(target)/$(mode)/img
-accel := on
+accel ?= on
 tcp ?= off
 
 sysroot := $(shell rustc --print sysroot)
@@ -18,11 +18,6 @@ ifeq ($(tcp), on)
 	BUILD_ARGS += --features tcp
 else
 	BUILD_ARGS += --no-default-features
-endif
-
-VSOCK_BUILD_ARGS =
-ifeq ($(mode), release)
-	VSOCK_BUILD_ARGS += --release
 endif
 
 QEMU_ARGS += \
